project-pgkram/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
cors==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Azure AI Inference SDK
azure-ai-inference==1.0.0b7
azure-core==1.30.2


